requirements.txt
fastapi
uvicorn
gunicorn
botbuilder-core
botbuilder-schema
aiohttp
azure-cli
langchain
langchain-community
langchain-core
langchain-openai
openai
pydantic
python-dotenv
python-pptx
redis
azure-storage-blob
jsonschema
pytest
pytest-asyncio
